+++ NEWS.md
@@ -2,5 +2,6 @@ exploreR 0.2.0.9000
 -------------------
 
 * added correlation heatmap for continuous and discrete features.
-* added NEWS.md
+* changed grid layout for `BoxplotContinuous`, `BarDiscrete` and `HistogramContinuous`.
+* added `NEWS.md`.